split files
Split the files, so the speakercraft control has its own file.

also setup as platform rather then individual components so that it can all be added with a single entry in yaml.
